Use compose-gstring-for-graphic instead of font-shape-gstring
Apparently, compose-gstring-for-graphic is higher level and able to
submit characters that have no ligature in the current font for shaping
by the font engine, which will just gracefully handle it with existing
glyphs. The result is that Emacs can treat the adjacent characters as a
single glyph even if it is rendered with two separate glyphs. With this
change, the composition table can have a superset of the ligatures in
the current font.