Remove macOS modifier key remapping
When working with an external keyboard these keys are in whatever place
is normal for the keyboard. More likely than not, that keyboard is
either a PC keyboard or a specialist keyboard with non-standard modifier
key locations.

In the event that the keyboard is a standard Apple keyboard, swapping
these at runtime is easy enough.