Remove mpd configuration
Running an mpd service as the login user allows it to output audio to
the audio server that runs as the login user, such as pulseaudio or
pipewire. However, it requires disabling the system mpd and enabling the
user mpd, which is the opposite of the Debian / Ubuntu default for the
package.

With the music on its own partition or dataset, it makes sense to let a
system mpd manage the library. It can run an httpd output for clients to
connect.