Make explicit use of variable-pitch-mode use the frame default
Now that the frame default font is variable pitch and fixed-pitch-mode
is in use to set buffers to explicitly use a fixed pitch face, make the
variable pitch face inherit from (now variable pitch) default face.

This change fixes font display for modes that mix fixed pitch faces with
variable pitch faces in the same buffer, which assume that the default
face is fixed pitch.