Fix symbol font configuration
The nerd font symbols use the private use area, not just the symbols
area, so set those as the fontset font for all of unicode, and prepend
it to the fontset. Prepending should be safe here because the font only
contains symbols, so regular glyphs will use the regular font.

Remove the unnecessary conditional that sets fonts only when running in
a windowing system.