Revert "Add /usr/local/bin to path explicitly"
It is possible that /usr/local/bin should not be added to the path
except by login shells. For example, if something like a cronjob gets
invoked as the user outside of a desktop environment context, then it
perhaps shouldn't be using things in /usr/local/bin. It's not entirely
clear that this would be a problem, but better to trust the judgement of
Apple and Ubuntu and whomever else that has /usr/local/bin set up only
through zprofile.

This reverts commit ce68caa0779da01b8a2867a5808d868b32547f8c.